+++ handlers/changeNotificationHandler.go
@@ -50,6 +50,7 @@ type ChangeNotificationHandler struct {
 	adminDB            *datastore.AdminServiceDatastore
 	tenantDB           *datastore.TenantServiceDatastore
 	metricsDB          datastore.DruidDatastore
+	batchSize          int64
 }
 
 // ChangeNotificationHandler singleton
@@ -79,6 +80,12 @@ func CreateChangeNotificationHandler() *ChangeNotificationHandler {
 		return nil
 	}
 
+	batchSize := int64(1000)
+	cfgBatchSize := cfg.GetInt(gather.CK_server_datastore_batchsize.String())
+	if cfgBatchSize > 0 {
+		batchSize = int64(cfgBatchSize)
+	}
+
 	changeNotifH = ChangeNotificationHandler{
 		brokers:            []string{broker},
 		topic:              defaultKafkaTopic,
@@ -86,6 +93,7 @@ func CreateChangeNotificationHandler() *ChangeNotificationHandler {
 		adminDB:            &adminDB,
 		provisioningEvents: make(chan *ChangeEvent, 20),
 		metricsDB:          druid.NewDruidDatasctoreClient(),
+		batchSize:          batchSize,
 	}
 
 	//	go changeNotifH.readFromKafka(broker, defaultKafkaTopic)
@@ -277,7 +285,7 @@ func debugAddFakeMonitoredObjects() []*tenmod.MonitoredObject {
 }
 
 func (c *ChangeNotificationHandler) updateMetricsDatastoreMetadata(tenantID string) {
-	monitoredObjects, err := (*c.tenantDB).GetAllMonitoredObjects(tenantID)
+	monitoredObjects, err := c.getAllMonitoredObjects(tenantID)
 	if err != nil {
 		logger.Log.Error("Failed to get objects", err.Error())
 		return
@@ -296,6 +304,30 @@ func (c *ChangeNotificationHandler) updateMetricsDatastoreMetadata(tenantID stri
 		logger.Log.Infof("Updated metadata in metric DB for tenant %s", tenantID)
 	}
 
+}
+
+// getAllMonitoredObjects - uses the paginated DB call to acquire all monitored objects
+func (c *ChangeNotificationHandler) getAllMonitoredObjects(tenantID string) ([]*tenmod.MonitoredObject, error) {
+
+	result := make([]*tenmod.MonitoredObject, 0)
+	startKey := ""
+
+	for true {
+		monitoredObjects, paginationOffsets, err := (*c.tenantDB).GetAllMonitoredObjectsByPage(tenantID, startKey, c.batchSize)
+		if err != nil {
+			return nil, err
+		}
+
+		result = append(result, monitoredObjects...)
+
+		if len(paginationOffsets.Next) == 0 {
+			break
+		}
+
+		startKey = paginationOffsets.Next
+	}
+
+	return result, nil
 }
 
 func (c *ChangeNotificationHandler) pollChanges(lastSyncTimestamp int64, fullRefresh bool) error {