+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 ## Current Release 
+### 0.178.0 
+**Release Date:** Thu Jul 26 18:59:59 UTC 2018     
+## Previous Releases 
+### 0.177.0 
+**Release Date:** Thu Jul 26 15:18:15 UTC 2018     
+* Feature - pagination for v2 GetAllMonitoredObjects API
+### 0.176.0 
+**Release Date:** Wed Jul 25 14:09:56 UTC 2018     
+* Fix raw metrics api to include unclean by default
 ### 0.175.0 
 **Release Date:** Mon Jul 23 20:29:41 UTC 2018     
-## Previous Releases 
 ### 0.174.0 
 **Release Date:** Fri Jul 20 17:57:47 UTC 2018     
 ### 0.173.0